Fix toolchain location.
Apparently we've been packing an empty tarball of the toolchain since
I split up the toolchain build because the script defines the install
path in a few different places.

Fix the toolchain install path to match the format we expect
(toolchains/$HOST/$TOOLCHAIN/prebuilt), improve logging, and actually
fail the build if we can't package the toolchain.

Change-Id: I093ca747af13b5ed093cb43499ea1fdb9aed0430